+++ fluid/fluid.cxx
@@ -30,6 +30,7 @@
 
 #include "alignment_panel.h"
 #include "function_panel.h"
+#include "sourceview_panel.h"
 #include "template_panel.h"
 #include "about_panel.h"
 
@@ -347,13 +348,6 @@ bool confirm_project_clear() {
   return true;
 }
 
-// ---- Sourceview definition
-
-void update_sourceview_position();
-void update_sourceview_position_cb(Fl_Tabs*, void*);
-void update_sourceview_cb(Fl_Button*, void*);
-void update_sourceview_timer(void*);
-
 // ----
 
 extern Fl_Window *the_panel;
@@ -438,7 +432,7 @@ void leave_project_dir() {
  \param[in] X, Y, W, H default size and position if nothing is specified in the preferences
  \return 1 if the caller should make the window visible, 0 if hidden.
  */
-char position_window(Fl_Window *w, const char *prefsName, int Visible, int X, int Y, int W=0, int H=0 ) {
+char position_window(Fl_Window *w, const char *prefsName, int Visible, int X, int Y, int W, int H) {
   Fl_Preferences pos(fluid_prefs, prefsName);
   if (prevpos_button->value()) {
     pos.get("x", X, X);
@@ -942,7 +936,7 @@ bool merge_project_file(const Fl_String &filename_arg) {
     free((void *)filename);
     filename = oldfilename;
     if (main_window) set_modflag(modflag);
-    return;
+    return false;
   }
   undo_resume();
   widget_browser->rebuild();
@@ -958,6 +952,7 @@ bool merge_project_file(const Fl_String &filename_arg) {
     if (oldfilename) free((void *)oldfilename);
   }
   g_project.update_settings_dialog();
+  return true;
 }
 
 /**
@@ -1586,43 +1581,12 @@ void toggle_widgetbin_cb(Fl_Widget *, void *) {
   }
 }
 
-/**
- Show or hide the source code preview.
- The state is stored in the app preferences.
- */
 void toggle_sourceview_cb(Fl_Double_Window *, void *) {
-  if (!sourceview_panel) {
-    make_sourceview();
-    sourceview_panel->callback((Fl_Callback*)toggle_sourceview_cb);
-    Fl_Preferences svp(fluid_prefs, "sourceview");
-    int autorefresh;
-    svp.get("autorefresh", autorefresh, 1);
-    sv_autorefresh->value(autorefresh);
-    int autoposition;
-    svp.get("autoposition", autoposition, 1);
-    sv_autoposition->value(autoposition);
-    int tab;
-    svp.get("tab", tab, 0);
-    if (tab>=0 && tab<sv_tab->children()) sv_tab->value(sv_tab->child(tab));
-    if (!position_window(sourceview_panel,"sourceview_pos", 0, 320, 120, 550, 500)) return;
-  }
-
-  if (sourceview_panel->visible()) {
-    sourceview_panel->hide();
-    sourceview_item->label("Show Source Code...");
-  } else {
-    sourceview_panel->show();
-    sourceview_item->label("Hide Source Code...");
-    update_sourceview_cb(0,0);
-  }
+  sourceview_toggle_visibility();
 }
 
-/**
- Show or hide the source code preview, called from a button.
- The state is stored in the app preferences.
- */
 void toggle_sourceview_b_cb(Fl_Button*, void *) {
-  toggle_sourceview_cb(0,0);
+  sourceview_toggle_visibility();
 }
 
 /**
@@ -1823,149 +1787,7 @@ void set_modflag(int mf, int mfc) {
   }
   // if the UI was modified in any way, update the Source View panel
   if (sourceview_panel && sourceview_panel->visible() && sv_autorefresh->value())
-  {
-    // we will only update earliest 0.5 seconds after the last change, and only
-    // if no other change was made, so dragging a widget will not generate any
-    // CPU load
-    Fl::remove_timeout(update_sourceview_timer, 0);
-    Fl::add_timeout(0.5, update_sourceview_timer, 0);
-  }
-}
-
-// ---- Sourceview implementation
-
-static char *sv_source_filename = NULL;
-static char *sv_header_filename = NULL;
-static char *sv_design_filename = NULL;
-
-/**
- Update the header and source code highlighting depending on the
- currently selected object
-
- The Source View system offers an immediate preview of the code
- files that will be generated by FLUID. It also marks the code
- generated for the last selected item in the header and the source
- file.
- */
-void update_sourceview_position()
-{
-  if (!sourceview_panel || !sourceview_panel->visible())
-    return;
-  if (sv_autoposition->value()==0)
-    return;
-  if (sourceview_panel && sourceview_panel->visible() && Fl_Type::current) {
-    int pos0, pos1;
-    if (sv_source->visible_r()) {
-      pos0 = Fl_Type::current->code_position;
-      pos1 = Fl_Type::current->code_position_end;
-      if (pos0>=0) {
-        if (pos1<pos0)
-          pos1 = pos0;
-        sv_source->buffer()->highlight(pos0, pos1);
-        int line = sv_source->buffer()->count_lines(0, pos0);
-        sv_source->scroll(line, 0);
-      }
-    }
-    if (sv_header->visible_r()) {
-      pos0 = Fl_Type::current->header_position;
-      pos1 = Fl_Type::current->header_position_end;
-      if (pos0>=0) {
-        if (pos1<pos0)
-          pos1 = pos0;
-        sv_header->buffer()->highlight(pos0, pos1);
-        int line = sv_header->buffer()->count_lines(0, pos0);
-        sv_header->scroll(line, 0);
-      }
-    }
-  }
-}
-
-/**
- Callback to update the sourceview position.
- */
-void update_sourceview_position_cb(Fl_Tabs*, void*)
-{
-  // make sure that the selected tab shows the current view
-  update_sourceview_cb(0,0);
-  // highlight the selected widget in the selected tab
-  update_sourceview_position();
-}
-
-/**
- Generate a header, source, strings, or design file in a temporary directory
- and load those into the Code Viewer widgets.
- */
-void update_sourceview_cb(Fl_Button*, void*)
-{
-  if (!sourceview_panel || !sourceview_panel->visible())
-    return;
-
-  if (!sv_source_filename) {
-    sv_source_filename = (char*)malloc(FL_PATH_MAX);
-    fluid_prefs.getUserdataPath(sv_source_filename, FL_PATH_MAX);
-    strlcat(sv_source_filename, "source_view_tmp.cxx", FL_PATH_MAX);
-  }
-  if (!sv_header_filename) {
-    sv_header_filename = (char*)malloc(FL_PATH_MAX);
-    fluid_prefs.getUserdataPath(sv_header_filename, FL_PATH_MAX);
-    strlcat(sv_header_filename, "source_view_tmp.h", FL_PATH_MAX);
-  }
-  if (!sv_design_filename) {
-    sv_design_filename = (char*)malloc(FL_PATH_MAX);
-    fluid_prefs.getUserdataPath(sv_design_filename, FL_PATH_MAX);
-    strlcat(sv_design_filename, "source_view_tmp.fl", FL_PATH_MAX);
-  }
-
-  if (sv_project->visible_r()) {
-    write_file(sv_design_filename);
-    int top = sv_project->top_line();
-    sv_project->buffer()->loadfile(sv_design_filename);
-    sv_project->scroll(top, 0);
-  } else if (sv_strings->visible_r()) {
-    static const char *exts[] = { ".txt", ".po", ".msg" };
-    char fn[FL_PATH_MAX];
-    fluid_prefs.getUserdataPath(fn, FL_PATH_MAX);
-    fl_filename_setext(fn, FL_PATH_MAX, exts[g_project.i18n_type]);
-    write_strings(fn);
-    int top = sv_strings->top_line();
-    sv_strings->buffer()->loadfile(fn);
-    sv_strings->scroll(top, 0);
-  } else if (sv_source->visible_r() || sv_header->visible_r()) {
-    g_project.basename = fl_filename_name(sv_source_filename);
-    g_project.basename = fl_filename_setext(g_project.basename, "");
-    Fl_String code_file_name_bak = g_project.code_file_name;
-    g_project.code_file_name = sv_source_filename;
-    Fl_String header_file_name_bak = g_project.header_file_name;
-    g_project.header_file_name = sv_header_filename;
-
-    // generate the code and load the files
-    Fd_Code_Writer f;
-    // generate files
-    if (f.write_code(sv_source_filename, sv_header_filename, true))
-    {
-      // load file into source editor
-      int pos = sv_source->top_line();
-      sv_source->buffer()->loadfile(sv_source_filename);
-      sv_source->scroll(pos, 0);
-      // load file into header editor
-      pos = sv_header->top_line();
-      sv_header->buffer()->loadfile(sv_header_filename);
-      sv_header->scroll(pos, 0);
-      // update the source code highlighting
-      update_sourceview_position();
-    }
-
-    g_project.code_file_name = code_file_name_bak;
-    g_project.header_file_name = header_file_name_bak;
-  }
-}
-
-/**
- This is called by the timer itself
- */
-void update_sourceview_timer(void*)
-{
-  update_sourceview_cb(0,0);
+    sourceview_defer_update();
 }
 
 // ---- Main program entry point
