+++ fluid/function_panel.cxx
@@ -1479,142 +1479,4 @@ Fl_Window* make_widgetbin() {
   return widgetbin_panel;
 }
 
-Fl_Double_Window *sourceview_panel=(Fl_Double_Window *)0;
-
-Fl_Tabs *sv_tab=(Fl_Tabs *)0;
-
-CodeViewer *sv_source=(CodeViewer *)0;
-
-CodeViewer *sv_header=(CodeViewer *)0;
-
-TextViewer *sv_strings=(TextViewer *)0;
-
-TextViewer *sv_project=(TextViewer *)0;
-
-Fl_Light_Button *sv_autorefresh=(Fl_Light_Button *)0;
-
-Fl_Light_Button *sv_autoposition=(Fl_Light_Button *)0;
-
-Fl_Double_Window* make_sourceview() {
-  { sourceview_panel = new Fl_Double_Window(520, 490, "Code View");
-    sourceview_panel->callback((Fl_Callback*)toggle_sourceview_cb);
-    sourceview_panel->align(Fl_Align(FL_ALIGN_CLIP|FL_ALIGN_INSIDE));
-    { sv_tab = new Fl_Tabs(10, 10, 500, 440);
-      sv_tab->selection_color((Fl_Color)4);
-      sv_tab->labelcolor(FL_BACKGROUND2_COLOR);
-      sv_tab->callback((Fl_Callback*)update_sourceview_position_cb);
-      { Fl_Group* o = new Fl_Group(10, 35, 500, 415, "Source");
-        o->labelsize(13);
-        { CodeViewer* o = sv_source = new CodeViewer(20, 50, 480, 390);
-          sv_source->box(FL_DOWN_FRAME);
-          sv_source->color(FL_BACKGROUND2_COLOR);
-          sv_source->selection_color(FL_SELECTION_COLOR);
-          sv_source->labeltype(FL_NORMAL_LABEL);
-          sv_source->labelfont(0);
-          sv_source->labelsize(14);
-          sv_source->labelcolor(FL_FOREGROUND_COLOR);
-          sv_source->textfont(4);
-          sv_source->textsize(11);
-          sv_source->align(Fl_Align(FL_ALIGN_TOP));
-          sv_source->when(FL_WHEN_RELEASE);
-          Fl_Group::current()->resizable(sv_source);
-          o->linenumber_width(60);
-          o->linenumber_size(o->Fl_Text_Display::textsize());
-        } // CodeViewer* sv_source
-        o->end();
-        Fl_Group::current()->resizable(o);
-      } // Fl_Group* o
-      { Fl_Group* o = new Fl_Group(10, 35, 500, 415, "Header");
-        o->labelsize(13);
-        o->hide();
-        { CodeViewer* o = sv_header = new CodeViewer(20, 50, 480, 390);
-          sv_header->box(FL_DOWN_FRAME);
-          sv_header->color(FL_BACKGROUND2_COLOR);
-          sv_header->selection_color(FL_SELECTION_COLOR);
-          sv_header->labeltype(FL_NORMAL_LABEL);
-          sv_header->labelfont(0);
-          sv_header->labelsize(14);
-          sv_header->labelcolor(FL_FOREGROUND_COLOR);
-          sv_header->textfont(4);
-          sv_header->textsize(11);
-          sv_header->align(Fl_Align(FL_ALIGN_TOP));
-          sv_header->when(FL_WHEN_RELEASE);
-          Fl_Group::current()->resizable(sv_header);
-          o->linenumber_width(60);
-          o->linenumber_size(o->Fl_Text_Display::textsize());
-        } // CodeViewer* sv_header
-        o->end();
-      } // Fl_Group* o
-      { Fl_Group* o = new Fl_Group(10, 35, 500, 415, "Strings");
-        o->labelsize(13);
-        o->hide();
-        { TextViewer* o = sv_strings = new TextViewer(20, 50, 480, 390);
-          sv_strings->box(FL_DOWN_FRAME);
-          sv_strings->color(FL_BACKGROUND2_COLOR);
-          sv_strings->selection_color(FL_SELECTION_COLOR);
-          sv_strings->labeltype(FL_NORMAL_LABEL);
-          sv_strings->labelfont(0);
-          sv_strings->labelsize(14);
-          sv_strings->labelcolor(FL_FOREGROUND_COLOR);
-          sv_strings->textfont(4);
-          sv_strings->textsize(11);
-          sv_strings->align(Fl_Align(FL_ALIGN_TOP));
-          sv_strings->when(FL_WHEN_RELEASE);
-          Fl_Group::current()->resizable(sv_strings);
-          o->linenumber_width(60);
-          o->linenumber_size(o->Fl_Text_Display::textsize());
-        } // TextViewer* sv_strings
-        o->end();
-      } // Fl_Group* o
-      { Fl_Group* o = new Fl_Group(10, 35, 500, 415, "Project");
-        o->labelsize(13);
-        o->hide();
-        { TextViewer* o = sv_project = new TextViewer(20, 50, 480, 390);
-          sv_project->box(FL_DOWN_FRAME);
-          sv_project->color(FL_BACKGROUND2_COLOR);
-          sv_project->selection_color(FL_SELECTION_COLOR);
-          sv_project->labeltype(FL_NORMAL_LABEL);
-          sv_project->labelfont(0);
-          sv_project->labelsize(14);
-          sv_project->labelcolor(FL_FOREGROUND_COLOR);
-          sv_project->textfont(4);
-          sv_project->textsize(11);
-          sv_project->align(Fl_Align(FL_ALIGN_TOP));
-          sv_project->when(FL_WHEN_RELEASE);
-          Fl_Group::current()->resizable(sv_project);
-          o->linenumber_width(60);
-          o->linenumber_size(o->Fl_Text_Display::textsize());
-        } // TextViewer* sv_project
-        o->end();
-      } // Fl_Group* o
-      sv_tab->end();
-      Fl_Group::current()->resizable(sv_tab);
-    } // Fl_Tabs* sv_tab
-    { Fl_Group* o = new Fl_Group(10, 460, 500, 20);
-      { Fl_Button* o = new Fl_Button(10, 460, 61, 20, "Refresh");
-        o->labelsize(11);
-        o->callback((Fl_Callback*)update_sourceview_cb);
-      } // Fl_Button* o
-      { Fl_Light_Button* o = sv_autorefresh = new Fl_Light_Button(76, 460, 91, 20, "Auto-Refresh");
-        sv_autorefresh->labelsize(11);
-        o->callback((Fl_Callback*)update_sourceview_cb);
-      } // Fl_Light_Button* sv_autorefresh
-      { sv_autoposition = new Fl_Light_Button(172, 460, 89, 20, "Auto-Position");
-        sv_autoposition->labelsize(11);
-      } // Fl_Light_Button* sv_autoposition
-      { Fl_Button* o = new Fl_Button(460, 460, 50, 20, "Close");
-        o->labelsize(11);
-        o->callback((Fl_Callback*)toggle_sourceview_b_cb);
-      } // Fl_Button* o
-      { Fl_Box* o = new Fl_Box(265, 460, 190, 20);
-        Fl_Group::current()->resizable(o);
-      } // Fl_Box* o
-      o->end();
-    } // Fl_Group* o
-    sourceview_panel->size_range(384, 120);
-    sourceview_panel->end();
-  } // Fl_Double_Window* sourceview_panel
-  return sourceview_panel;
-}
-
 //
